# Assignment/MVCCore_OAS3Client/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebApi1",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Category": {
      "get": {
        "tags": [
          "Category"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Category"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Category/{id}": {
      "get": {
        "tags": [
          "Category"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Category"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Category"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Product": {
      "get": {
        "tags": [
          "Product"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Product"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/GetProductSRC/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Category": {
        "type": "object",
        "properties": {
          "CategoryId": {
            "type": "integer",
            "format": "int32"
          },
          "CategoryName": {
            "type": "string",
            "nullable": true
          },
          "BasePrice": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "SubCategories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubCategory"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Manufracturer": {
        "type": "object",
        "properties": {
          "ManufractureId": {
            "type": "integer",
            "format": "int32"
          },
          "ManufractureName": {
            "type": "string",
            "nullable": true
          },
          "ManufractureAdderess": {
            "type": "string",
            "nullable": true
          },
          "ManufractureCity": {
            "type": "string",
            "nullable": true
          },
          "ManufractureState": {
            "type": "string",
            "nullable": true
          },
          "Products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Product": {
        "type": "object",
        "properties": {
          "ProductUniqueId": {
            "type": "integer",
            "format": "int32"
          },
          "ProductId": {
            "type": "string",
            "nullable": true
          },
          "ProductName": {
            "type": "string",
            "nullable": true
          },
          "Descrition": {
            "type": "string",
            "nullable": true
          },
          "Price": {
            "type": "number",
            "format": "double"
          },
          "SubCategoryId": {
            "type": "integer",
            "format": "int32"
          },
          "ManufractureId": {
            "type": "integer",
            "format": "int32"
          },
          "Manufracture": {
            "$ref": "#/components/schemas/Manufracturer"
          },
          "SubCategory": {
            "$ref": "#/components/schemas/SubCategory"
          }
        },
        "additionalProperties": false
      },
...
      "SubCategory": {
        "type": "object",
        "properties": {
          "SubCategoryId": {
            "type": "integer",
            "format": "int32"
          },
          "SubCategoryName": {
            "type": "string",
            "nullable": true
          },
          "CategoryId": {
            "type": "integer",
            "format": "int32"
          },
          "Category": {
            "$ref": "#/components/schemas/Category"
          },
          "Products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}